Split CI run summary aggregation helpers

Copilot-Session: daf71962-ce66-4366-8ba6-700a2f0829bf
